[refactor] Move DummyModel to tests
- Deleted the `DummyModel` class and its configuration from modeling.
- Moved the `DummyModel` class to tests.
