+++ tensorrt_llm/_torch/models/__init__.py
@@ -4,7 +4,6 @@ from .modeling_auto import AutoModelForCausalLM
 from .modeling_bert import BertForSequenceClassification
 from .modeling_clip import CLIPVisionModel
 from .modeling_deepseekv3 import DeepseekV3ForCausalLM
-from .modeling_dummy import DummyModel
 from .modeling_exaone4 import Exaone4ForCausalLM
 from .modeling_gemma3 import Gemma3ForCausalLM
 from .modeling_gemma3vl import Gemma3VLM
@@ -40,7 +39,6 @@ __all__ = [
     "BertForSequenceClassification",
     "CLIPVisionModel",
     "DeepseekV3ForCausalLM",
-    "DummyModel",
     "Exaone4ForCausalLM",
     "Gemma3ForCausalLM",
     "Gemma3VLM",
